+++ keys.go
@@ -22,6 +22,24 @@ package clortho
 // that have a keyID.
 type Keys []Key
 
+// AppendKeyIDs appends the key Id of each key to the supplied slice,
+// then returns the result.
+func (ks Keys) AppendKeyIDs(v []string) []string {
+	if cap(v) < len(v)+len(ks) {
+		// reduce the number of allocations
+		v = append(
+			make([]string, 0, len(v)+len(ks)),
+			v...,
+		)
+	}
+
+	for _, k := range ks {
+		v = append(v, k.KeyID())
+	}
+
+	return v
+}
+
 // Len returns the count of Key instances in this collection.
 func (ks Keys) Len() int {
 	return len(ks)