+++ parser_test.go
@@ -24,7 +24,7 @@ import (
 	"errors"
 	"testing"
 
-	"github.com/lestrrat-go/jwx/jwa"
+	"github.com/lestrrat-go/jwx/v2/jwa"
 	"github.com/stretchr/testify/suite"
 )